Problem Statement
Design and code a responsive webpage which is a graph profile of a
string of text. You
need to have two input boxes in your UI design, where user can key-in
his name and a
string of text. The parent node of the graph will represent the given
name. The other
connected nodes of the graph would be alphabets of the given string; the
more
number of times an alphabet occurs in the string; the size of the node
should represent
the same (i.e. higher the occurrence bigger the child nodes). Ensure
each alphabet is
shown in a different colour node (circle) dynamically.